Mostly finish extracting common HTML into a Jekyll layout.
This uses the 'default' layout for mostly everything, including the new
index page for the resources category.

Making the breadcrumbs work was a bit difficult and still much to be
desired, but it works well enough for a prototype so off we go!

Everything is still static at this point, the next fun thing to do would
probably be to run an API backend somewhere and show off trying to do
something dynamic using client-side JS.

Or obviously, continuing to populate static content.