mobile: Move main content upwards.
The sidebar being on the left has been done up to now by making its
column the first one seen in the main HTML, followed by the "real"
content.

This results in the sidebar being the topmost element on mobile though,
which is near-useless.

I had experimented with various ways of trying to ensure the sidebar
stayed on the left on desktop versions, but actually I think it looks
better this way anyways.